Moved the trait object generation away from the character definition
Up until now, a Trait object was created when a character was defined.
This, however, meant that different characters could have different
trait sets, but also that the whole specification for a trait had to be
done when making the character all over again.
This commit moved the trait object creation away from the character
definition into a seperate script file, and every character now gets her
own copy of the trait definition (a deepcopy of a global trait
dictionary).