Goals and events
Started to implement goals. Saw that I would have to have much more
mighty events to be able to implement goals in a satisfying way. So this
was mainly an events commit. Events now can apply effects to the
characters in the event, although that is not tested very thoroughly
yet.